requirements.txt
streamlit
weaviate-client
requests
pandas
Pillow
